Rename controller.js to messagecontroller.js for better clarity
